+++ test/libsolidity/syntaxTests/types/rational_number_literal_unlimited_1.sol
@@ -1,9 +1,9 @@
 contract c {
     function bignum() public {
-        uint256 a;
+        uint a;
         a = 1e1233 / 1e1233; // 1e1233 is still fine
         a = 1e1234; // 1e1234 is too big
     }
 }
 // ----
-// TypeError: (128-134): Type int_const 1000...(1227 digits omitted)...0000 is not implicitly convertible to expected type uint256.
+// Warning: (17-159): Function state mutability can be restricted to pure
