+++ test/libsolidity/syntaxTests/types/rational_number_literal_unlimited_2.sol
@@ -1,9 +1,9 @@
 contract c {
     function bignum() public {
-        uint256 a;
+        uint a;
         a = 134562324532464234452335168163516E1200 / 134562324532464234452335168163516E1200; // still fine
         a = 1345623245324642344523351681635168E1200; // too large
     }
 }
 // ----
-// TypeError: (182-221): Type int_const 1345...(1226 digits omitted)...0000 is not implicitly convertible to expected type uint256.
+// Warning: (17-238): Function state mutability can be restricted to pure
