+++ test/libsolidity/syntaxTests/types/rational_number_literal_unlimited_3.sol
@@ -1,9 +1,9 @@
 contract c {
     function bignum() public {
-        uint256 a;
+        uint a;
         a = 134562324532464.234452335168163517E1200 / 134562324532464.234452335168163517E1200; // still fine
         a = 134562324532464.2344523351681635177E1200; // too large
     }
 }
 // ----
-// TypeError: (184-224): Type int_const 1345...(1207 digits omitted)...0000 is not implicitly convertible to expected type uint256.
+// Warning: (17-241): Function state mutability can be restricted to pure
