+++ test/libsolidity/syntaxTests/types/rational_number_mul_limit.sol
@@ -1,9 +1,8 @@
 contract c {
     function f() public pure {
-        int a;
+        int256 a;
         a = (1<<4095)*(1<<4095);
     }
 }
 // ----
-// TypeError: (71-90): Operator * not compatible with types int_const 5221...(1225 digits omitted)...5168 and int_const 5221...(1225 digits omitted)...5168
-// TypeError: (71-90): Type int_const 5221...(1225 digits omitted)...5168 is not implicitly convertible to expected type int256.
+// TypeError: (74-93): Type uint is not implicitly convertible to expected type int256.
